Address PR #79 bot review round 2 (3 more nits)
- PerCardEnricher extracted to its own file, matching the
  one-type-per-file convention used elsewhere in the project.
- IEnricher.cs is now just the interface declaration.
- PerCardEnricher XML doc now notes that the reverse-iteration
  side-effect appends issues in reverse row order within a pass —
  consumers should sort by ImportIssue.RowNumber if order matters.
- CardmarketIdEnricher: comment on the error path clarifies that
  unresolved duplicate-ID inputs produce one issue per affected row
  (not one per distinct ID), preserving row-level traceability.
