+++ MtgCsvHelper/Enrichment/CardmarketIdEnricher.cs
@@ -50,6 +50,8 @@ public sealed class CardmarketIdEnricher(ICardmarketResolver resolver) : IEnrich
 			{
 				// Without a name/set, we can't write a meaningful row — drop the card.
 				// This is data loss (Error), not just degraded fidelity (Warning).
+				// One error per affected row (not per distinct ID), so duplicate-ID inputs that
+				// fail to resolve produce one issue per row — keeps row-level traceability.
 				issues.Add(new ImportIssue(IssueSeverity.Error, row.RowNumber, $"Cardmarket ID {id} not found in Scryfall data — card skipped"));
 				rows.RemoveAt(i);
 			}
